backend/pyproject.toml
[tool.poetry]
name = "bucket-interior"
version = "0.1.0"
description = "bucket-interior"
authors = ["Yewon Kim <yewon.devv@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
fastapi = "^0.70.0"
uvicorn = "^0.17.6"
torch = "^1.9.1"
SQLAlchemy = "^1.4.36"
PyMySQL = "^1.0.2"
pydantic = {extras = ["email"], version = "^1.9.1"}
gcsfs = "^2022.5.0"
google-cloud-storage = "^2.3.0"
numpy = "^1.22.4"
python-multipart = "^0.0.5"
PyYAML = "^6.0"
tqdm = "^4.64.0"
image = "^1.5.33"
tensorboard = "2.4.1"
matplotlib = "3.3.4"
torchvision = "^0.10.1"
async_generator = "^1.10"
async-exit-stack = "^1.0.1"
mysql-connector-python-rf = "^2.2.2"
mysql = "^0.0.3"


[tool.poetry.dev-dependencies]
pytest = "^7.1.2"


[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
